+++ common/src/main/java/org/geogebra/common/export/pstricks/GeoGebraToPstricks.java
@@ -472,9 +472,9 @@ public abstract class GeoGebraToPstricks extends GeoGebraExport {
 		// get point on line g
 		((AlgoSlope) geo.getParentAlgorithm()).getInhomPointOnLine(coords);
 		// draw slope triangle
-		float x = (float) coords[0];
-		float y = (float) coords[1];
-		float xright = x + slopeTriangleSize;
+		double x = coords[0];
+		double y = coords[1];
+		double xright = x + slopeTriangleSize;
 		startBeamer(codeFilledObject);
 		codeFilledObject.append("\\pspolygon");
 		codeFilledObject.append(lineOptionCode(geo, true));
@@ -496,8 +496,8 @@ public abstract class GeoGebraToPstricks extends GeoGebraExport {
 		codeFilledObject.append("\n");
 		endBeamer(codeFilledObject);
 		// draw Label
-		float xLabelHor = (x + xright) / 2;
-		float yLabelHor = y - (float) ((euclidianView.getFont().getSize() + 2)
+		double xLabelHor = (x + xright) / 2;
+		double yLabelHor = y - ((euclidianView.getFont().getSize() + 2)
 				/ euclidianView.getYscale());
 		GColor geocolor = geo.getObjectColor();
 		startBeamer(codePoint);
